requirements.txt
streamlit
joblib
numpy
scikit-learn
mysql-connector-python